Add files via upload
Write a program that will allow a student to enter their name and then ask them to solve  

10 mathematical equations. The program should display two random numbers that are to  

be added, such as: 

247 

+ 129 

The program should allow the student to enter the answer. The program should then  

display whether the answer was right or wrong, and accumulate the correct values. After  

the 10 questions are asked, calculate the average correct. Then display the student name,  

the number correct, and the average correct in both decimal and percentage format. 

In addition to any system functions you may use, you might consider the following  

functions: 

A function that allows the student to enter their name. 

A function that gets two random numbers, anywhere from 1 to 500. 

A function that displays the equation and asks the user to enter their answer. 

A function that checks to see if the answer is right and accumulates the number  

right. 

A function that calculates the results. 

A function that displays the student name, the number right, and the average right. 

Your sample output might look as follows (random numbers will be different): 

Enter Student Name: Katie 

What is the answer to the following equation 

424 

+ 

28 

What is the sum: 472 

Wrong